+++ changelog.md
@@ -1,3 +1,9 @@
+## Next
+
+- Environment file is only changed when its content changes. This will improve deploy speed when `longEnvVars` is true.
+- Fix environment sometimes not ready when trying to migrate to using environment variables stored in s3
+- Fix `mup status` command 
+
 ## 0.6.2 - November 15, 2018
 
 - Fix `Throttling: Rate exceeded` errors